feat: configure database

+++ src/movies/movies.service.ts
@@ -1,4 +1,8 @@
-import { Injectable } from '@nestjs/common';
+import {
+  BadRequestException,
+  Injectable,
+  NotFoundException,
+} from '@nestjs/common';
 import { CreateMovieDto } from './dto/create-movie.dto';
 import { UpdateMovieDto } from './dto/update-movie.dto';
 import { InjectRepository } from '@nestjs/typeorm';
@@ -10,23 +14,41 @@ export class MoviesService {
   constructor(
     @InjectRepository(Movie) private movieRepository: Repository<Movie>,
   ) {}
-  create(createMovieDto: CreateMovieDto) {
-    return this.movieRepository.create(createMovieDto);
+  async create(createMovieDto: CreateMovieDto) {
+    try {
+      const newMovie = this.movieRepository.create(createMovieDto);
+      const { id } = await this.movieRepository.save(newMovie);
+      return { id };
+    } catch (e) {
+      throw new BadRequestException('Movie not saved');
+    }
   }
 
   findAll() {
     return this.movieRepository.find();
   }
 
-  findOne(id: number) {
-    return this.movieRepository.findOneBy({ id });
+  async findOne(id: number) {
+    const movie = await this.movieRepository.findOneBy({ id });
+    if (!movie) {
+      throw new NotFoundException('Movie not found');
+    }
+    return movie;
   }
 
-  update(id: number, updateMovieDto: UpdateMovieDto) {
+  async update(id: number, updateMovieDto: UpdateMovieDto) {
+    const movie = await this.movieRepository.findOneBy({ id });
+    if (!movie) {
+      throw new NotFoundException('Movie not found');
+    }
     return this.movieRepository.update(id, updateMovieDto);
   }
 
-  remove(id: number) {
+  async remove(id: number) {
+    const movie = await this.movieRepository.findOneBy({ id });
+    if (!movie) {
+      throw new NotFoundException('Movie not found');
+    }
     return this.movieRepository.delete(id);
   }
 }
